feature(logger-service): PR comments and some refactors.
